+++ btclib/tests/test_curves.py
@@ -9,31 +9,33 @@
 # or distributed except according to the terms contained in the LICENSE file.
 
 import unittest
+from typing import Dict
 
 from btclib.alias import INF, INFJ
 from btclib.curve import Curve, _jac_from_aff
 from btclib.curvemult import mult
-from btclib.curves import (BP_curves, NIST_curves, SEC2V1_curves, secp112r1,
-                           secp160r1, secp256k1)
+from btclib.curves import CURVES
 from btclib.numbertheory import mod_sqrt
 from btclib.secpoint import bytes_from_point, point_from_octets
 
 # test curves: very low cardinality
-ec13_11 = Curve(13, 7, 6, (1, 1), 11, 1, False)
-ec13_19 = Curve(13, 0, 2, (1, 9), 19, 1, False)
-ec17_13 = Curve(17, 6, 8, (0, 12), 13, 2, False)
-ec17_23 = Curve(17, 3, 5, (1, 14), 23, 1, False)
-ec19_13 = Curve(19, 0, 2, (4, 16), 13, 2, False)
-ec19_23 = Curve(19, 2, 9, (0, 16), 23, 1, False)
-ec23_19 = Curve(23, 9, 7, (5, 4), 19, 1, False)
-ec23_31 = Curve(23, 5, 1, (0, 1), 31, 1, False)
-low_card_curves = [ec13_11, ec13_19,  # 13 % 4 = 1; 13 % 8 = 5
-                   ec17_13, ec17_23,  # 17 % 4 = 1; 17 % 8 = 1
-                   ec19_13, ec19_23,  # 19 % 4 = 3; 19 % 8 = 3
-                   ec23_19, ec23_31   # 23 % 4 = 3; 23 % 8 = 7
-                   ]
+low_card_curves: Dict[str, Curve] = {}
+# 13 % 4 = 1; 13 % 8 = 5
+low_card_curves['ec13_11'] = Curve(13, 7, 6, (1, 1), 11, 1, False)
+low_card_curves['ec13_19'] = Curve(13, 0, 2, (1, 9), 19, 1, False)
+# 17 % 4 = 1; 17 % 8 = 1
+low_card_curves['ec17_13'] = Curve(17, 6, 8, (0, 12), 13, 2, False)
+low_card_curves['ec17_23'] = Curve(17, 3, 5, (1, 14), 23, 1, False)
+# 19 % 4 = 3; 19 % 8 = 3
+low_card_curves['ec19_13'] = Curve(19, 0, 2, (4, 16), 13, 2, False)
+low_card_curves['ec19_23'] = Curve(19, 2, 9, (0, 16), 23, 1, False)
+# 23 % 4 = 3; 23 % 8 = 7
+low_card_curves['ec23_19'] = Curve(23, 9, 7, (5, 4), 19, 1, False)
+low_card_curves['ec23_31'] = Curve(23, 5, 1, (0, 1), 31, 1, False)
 
-all_curves = low_card_curves + SEC2V1_curves + NIST_curves + BP_curves
+all_curves: Dict[str, Curve] = {}
+all_curves.update(low_card_curves)
+all_curves.update(CURVES)
 
 
 class TestEllipticCurve(unittest.TestCase):
@@ -82,7 +84,8 @@ class TestEllipticCurve(unittest.TestCase):
         self.assertRaises(UserWarning, Curve, 11, 2, 7, (6, 9), 7, 2, True)
 
         # x-coordinate not in [0, p-1]
-        self.assertRaises(ValueError, secp256k1.y, secp256k1.p)
+        ec = CURVES['secp256k1']
+        self.assertRaises(ValueError, ec.y, ec.p)
         # secp256k1.y(secp256k1.p)
 
         # INF point does not generate a prime order subgroup
@@ -90,7 +93,7 @@ class TestEllipticCurve(unittest.TestCase):
         # Curve(11, 2, 7, INF, 7, 2, 0, False)
 
     def test_all_curves(self):
-        for ec in all_curves:
+        for ec in all_curves.values():
             self.assertEqual(mult(0, ec.G, ec), INF)
             self.assertEqual(mult(0, ec.G, ec), INF)
 
@@ -130,13 +133,13 @@ class TestEllipticCurve(unittest.TestCase):
             self.assertEqual(mult(25, INF, ec), INF)
 
             ec_repr = repr(ec)
-            if ec in low_card_curves or ec.psize < 24:
+            if ec in low_card_curves.values() or ec.psize < 24:
                 ec_repr = ec_repr[:-1] + ", False)"
             ec2 = eval(ec_repr)
             self.assertEqual(str(ec), str(ec2))
 
     def test_octets2point(self):
-        for ec in all_curves:
+        for ec in all_curves.values():
             Q = mult(ec.p, ec.G, ec)  # just a random point, not INF
 
             Q_bytes = b'\x03' if Q[1] & 1 else b'\x02'
@@ -174,7 +177,7 @@ class TestEllipticCurve(unittest.TestCase):
             self.assertRaises(ValueError, point_from_octets, Q_bytes, ec)
 
         # invalid x coordinate
-        ec = secp256k1
+        ec = CURVES['secp256k1']
         x = 0xEEFDEA4CDB677750A420FEE807EACF21EB9898AE79B9768766E4FAA04A2D4A34
         xstr = format(x, '32X')
         self.assertRaises(ValueError, point_from_octets, "03" + xstr, ec)
@@ -191,7 +194,7 @@ class TestEllipticCurve(unittest.TestCase):
         self.assertRaises(ValueError, ec.is_on_curve, P)
 
     def test_opposite(self):
-        for ec in all_curves:
+        for ec in all_curves.values():
             Q = mult(ec.p, ec.G, ec)  # just a random point, not INF
             minus_Q = ec.negate(Q)
             self.assertEqual(ec.add(Q, minus_Q), INF)
@@ -206,7 +209,7 @@ class TestEllipticCurve(unittest.TestCase):
 
     def test_symmetry(self):
         """Methods to break simmetry: quadratic residue, odd/even, low/high"""
-        for ec in low_card_curves:
+        for ec in low_card_curves.values():
 
             # setup phase
             # compute quadratic residues
@@ -284,7 +287,7 @@ class TestEllipticCurve(unittest.TestCase):
         self.assertRaises(ValueError, ec.y_quadratic_residue, x, 2)
 
     def test_aff_jac_conversions(self):
-        for ec in all_curves:
+        for ec in all_curves.values():
             Q = mult(ec.p, ec.G, ec)  # just a random point, not INF
             QJ = _jac_from_aff(Q)
             checkQ = ec._aff_from_jac(QJ)
@@ -300,7 +303,7 @@ class TestEllipticCurve(unittest.TestCase):
             self.assertRaises(ValueError, ec.has_square_y, "Not a Point")
 
     def test_add(self):
-        for ec in all_curves:
+        for ec in all_curves.values():
             Q1 = mult(ec.p, ec.G, ec)  # just a random point, not INF
             Q1J = _jac_from_aff(Q1)